requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.5
django-admin-rangefilter==0.5.4
django-auth-ldap==2.1.1
django-bootstrap3==12.1.0
django-filter==2.2.0
django-import-export==2.0.2
django-paypal==1.0.0
django-render-block==0.6
django-templated-email==2.3.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
